create separate users module, fix config for find .env file
