+++ backend/package.json
@@ -31,6 +31,7 @@
     "@mikro-orm/postgresql": "~6.5.9",
     "@mikro-orm/sql-highlighter": "~1.0.1",
     "@nestjs/common": "^11.0.1",
+    "@nestjs/config": "^4.0.2",
     "@nestjs/core": "^11.0.1",
     "@nestjs/platform-express": "^11.0.1",
     "@nestjs/platform-fastify": "^11.1.8",
