+++ backend/src/auth/auth.guard.ts
@@ -6,7 +6,7 @@ import {
 } from '@nestjs/common';
 import { FastifyRequest } from 'fastify';
 import { SupabaseService } from './supabase.service';
-import { UsersService } from './users.service';
+import { UsersService } from '../users/users.service';
 
 interface RequestWithUser extends FastifyRequest {
   user?: ReturnType<UsersService['formatUserResponse']>;
