+++ backend/src/auth/auth.module.ts
@@ -1,16 +1,14 @@
 import { Module } from '@nestjs/common';
-import { MikroOrmModule } from '@mikro-orm/nestjs';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 import { SupabaseService } from './supabase.service';
-import { UsersService } from './users.service';
 import { AuthGuard } from './auth.guard';
-import { User } from '../entities/user.entity';
+import { UsersModule } from '../users/users.module';
 
 @Module({
-  imports: [MikroOrmModule.forFeature([User])],
+  imports: [UsersModule],
   controllers: [AuthController],
-  providers: [AuthService, SupabaseService, UsersService, AuthGuard],
-  exports: [AuthService, SupabaseService, UsersService, AuthGuard],
+  providers: [AuthService, SupabaseService, AuthGuard],
+  exports: [AuthService, SupabaseService, AuthGuard],
 })
 export class AuthModule {}
