+++ backend/src/modules/orm.module.ts
@@ -1,8 +1,9 @@
 import { Module } from '@nestjs/common';
 import { MikroOrmModule } from '@mikro-orm/nestjs';
+import config from '../mikro-orm.config';
 
 @Module({
-  imports: [MikroOrmModule.forRoot(), MikroOrmModule.forFeature({})],
+  imports: [MikroOrmModule.forRoot(config), MikroOrmModule.forFeature({})],
   exports: [MikroOrmModule],
 })
 export class OrmModule {}
